+++ repository/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.0.2] - 2019-09-09
+
+* `CachedRepository` allowing multiple cache strategies.
+
+## [1.0.1] - 2019-09-09
+
+* Minor fixes allowing `fetch` stream subscription to be closed after first
+  event.
+
 ## [1.0.0] - 2019-09-06
 
 * Removed unnecessary dependencies `flutter`, `hive`, `path_provider`,